Data-Scientist-Capstone/requirements.txt
Flask==2.3.2
matplotlib==3.7.2
numpy==1.25.1
orjson==3.9.2
Pillow==10.0.0
randaugment==1.0.2
scikit_learn==1.3.0
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
